Added ViewModel logic for EmojiListViewController
