+++ Challange-1-iOS/Challange-1-iOS/Persistence/AvatarPersistence.swift
@@ -39,7 +39,6 @@ class AvatarPersistence {
             }
         }
       
-      
     }
     
     func fetchAvatarData(_ resulthandler: @escaping ([NSManagedObject]) -> Void) {
