+++ Challange-1-iOS/Challange-1-iOS/ViewControllers/EmojisListViewController.swift
@@ -9,20 +9,20 @@ import UIKit
 
 class EmojisListViewController: UIViewController, Coordinating {
     private var collectionView: UICollectionView
-    var emojiService: EmojiService?
+//    var emojiService: EmojiService?
     
     var coordinator: Coordinator?
-
+    
     var liveEmojiStorage: LiveEmojiStorage = .init()
     
     var emojisList: [Emoji] = []
     
-    var strong = MockedDataSource()
+    var viewModel: EmojiViewModel?
     
-//    lazy var collectionView: UICollectionView = {
-//        let v = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
-//        return v
-//    }()
+    //    lazy var collectionView: UICollectionView = {
+    //        let v = UICollectionView(frame: .zero, collectionViewLayout: UICollectionViewFlowLayout())
+    //        return v
+    //    }()
     
     init(){
         let layout = UICollectionViewFlowLayout()
@@ -30,7 +30,7 @@ class EmojisListViewController: UIViewController, Coordinating {
         
         layout.minimumLineSpacing = 8
         layout.minimumInteritemSpacing = 4
-
+        
         collectionView = .init(frame: .zero, collectionViewLayout: layout)
         
         super.init(nibName: nil, bundle: nil)
@@ -69,27 +69,23 @@ class EmojisListViewController: UIViewController, Coordinating {
     
     private func setUpCollectionView() {
         title = "Emojis List"
-
+        
         collectionView.register(EmojiCell.self, forCellWithReuseIdentifier: EmojiCell.reuseCellIdentifier)
-
+        
         collectionView.delegate = self
         collectionView.dataSource = self
-        }
+    }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        emojiService?.getEmojisList({ (result: Result<[Emoji], Error>) in
-            switch result {
-            case .success(let success):
-                self.emojisList = success
-                DispatchQueue.main.async() { [weak self] in
-                    self?.collectionView.reloadData()
-                }
-            case .failure(let failure):
-                print("Error: \(failure)")
+        viewModel?.emojisList.bind(listener: { [weak self] emojiArray in
+            guard let emojiArray = emojiArray else { return }
+            self?.emojisList = emojiArray
+            DispatchQueue.main.async() { [weak self] in
+                self?.collectionView.reloadData()
             }
         })
-        
+        viewModel?.getEmojis()        
     }
 }
 
@@ -107,11 +103,11 @@ extension EmojisListViewController: UICollectionViewDataSource {
         
         return countEmojis
     }
-
+    
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         
         let cell : EmojiCell = collectionView.dequeueReusableCell(for: indexPath)
-
+        
         let url = emojisList[indexPath.row].emojiUrl
         
         cell.setUpCell(url: url)
@@ -120,38 +116,16 @@ extension EmojisListViewController: UICollectionViewDataSource {
     }
 }
 
-
-class MockedDataSource: NSObject, UICollectionViewDataSource {
-    var mockedEmojis: MockEmojiStorage = .init()
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        
-        return mockedEmojis.emojis.count
-    }
-
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
-        let cell : EmojiCell = collectionView.dequeueReusableCell(for: indexPath)
-
-        let url = mockedEmojis.emojis[indexPath.row].emojiUrl
-        
-        cell.setUpCell(url: url)
-        
-        return cell
-    }
-}
-
-
 extension EmojisListViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView,
-                  layout collectionViewLayout: UICollectionViewLayout,
-                  insetForSectionAt section: Int) -> UIEdgeInsets {
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        insetForSectionAt section: Int) -> UIEdgeInsets {
         return UIEdgeInsets(top: 1.0, left: 8.0, bottom: 1.0, right: 8.0)
     }
-
+    
     func collectionView(_ collectionView: UICollectionView,
-                   layout collectionViewLayout: UICollectionViewLayout,
-                   sizeForItemAt indexPath: IndexPath) -> CGSize {
+                        layout collectionViewLayout: UICollectionViewLayout,
+                        sizeForItemAt indexPath: IndexPath) -> CGSize {
         
         let layout = collectionViewLayout as! UICollectionViewFlowLayout
         let widthPerItem = collectionView.frame.width / 3 - layout.minimumInteritemSpacing
